(redesign/ci-cd) - Widen stock quantity to decimal for weight-measured products
Delta $5/$6 quantity halves (frontend handoff). Reverses the prior integer-only stock decision by explicit owner call: on-hand stock quantity is now decimal(18,3) domain-wide so weight-measured products (e.g. 2.5 kg) can be stocked, sold, adjusted, and transferred.

Widened Quantity on WarehouseItem, OpeningStock, StockAdjustment, OrderLine, and TemplateItem (TransactionLine/TransferLine were already decimal); the MoveStockAsync tuple; the request lines and the response DTOs that carry a quantity or a running balance (movements, stock items, adjustments, order/template lines) plus the computed TotalUnits/TotalStock aggregates. Added a HasQuantityPrecision(18,3) EF helper and a non-destructive AlterColumn migration on the five tables.

The WAC and negative-stock formulas were already decimal, so only types changed there. Removed the explicit (int) casts in MovementService and TransferMappings that would have truncated a fractional quantity (the compiler cannot flag those). LowStockThreshold stays int (a threshold, not on-hand quantity).

Tests: new fractional coverage — WAC recompute on 2.5+1.5 kg = 47.5, a 2.5 kg inter-warehouse transfer, a fractional decrease with no truncation, and a negative-stock block against fractional stock on hand.

+++ src/Ombor.Application/Extensions/StockMovementExtensions.cs
@@ -45,7 +45,7 @@ internal static class StockMovementExtensions
         this IApplicationDbContext context,
         int warehouseId,
         StockMovement movement,
-        IEnumerable<(int ProductId, int Quantity, decimal UnitPrice)> rawLines)
+        IEnumerable<(int ProductId, decimal Quantity, decimal UnitPrice)> rawLines)
     {
         var lines = rawLines
             .GroupBy(x => x.ProductId)
