+++ src/Ombor.Contracts/Responses/Product/CreateProductResponse.cs
@@ -41,6 +41,6 @@ public sealed record CreateProductResponse(
     bool IsArchived,
     ProductImageDto[] Images,
     ProductWarehouseItemDto[] WarehouseItems,
-    int TotalStock,
+    decimal TotalStock,
     decimal? AverageCost,
     ProductPackagingDto? Packaging);
